# api/app/src/swagger/swagger_spec.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1",
        "title": "MethodFitness"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/client/addClient": {
            "post": {
                "x-name": "/client/addClient",
                "description": "add client",
                "operationId": "/client/addClient",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/updateClientInfo": {
            "post": {
                "x-name": "/client/updateClientInfo",
                "description": "update Client Info",
                "operationId": "/client/updateClientInfo",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateClientInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/updateClientSource": {
            "post": {
                "x-name": "/client/updateClientSource",
                "description": "update Client Source",
                "operationId": "/client/updateClientSource",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateClientSource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/updateClientContact": {
            "post": {
                "x-name": "/client/updateClientContact",
                "description": "update Client Contact",
                "operationId": "/client/updateClientContact",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateClientContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/updateClientAddress": {
            "post": {
                "x-name": "/client/updateClientAddress",
                "description": "update Client Address",
                "operationId": "/client/updateClientAddress",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateClientAddress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/getClient/{id}": {
            "get": {
                "x-name": "client",
                "description": "retrieve single client by id",
                "operationId": "client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The client id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/client/archiveClient": {
            "post": {
                "x-name": "/client/archiveClient",
                "description": "archive Client",
                "operationId": "/client/archiveClient",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/archiveClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "client": {
            "title": "client",
            "properties": {
                "contact": {
                    "title": "contact",
                    "properties": {
                        "firstName": {
                            "type": "string"
                        },
                        "lastName": {
                            "type": "string"
                        },
                        "address": {
                            "title": "address",
                            "properties": {
                                "street1": {
                                    "type": "string"
                                },
                                "street2": {
                                    "type": "string"
                                },
                                "city": {
                                    "type": "string"
                                },
                                "state": {
                                    "type": "string",
                                    "x-input": "select"
                                },
                                "zipCode": {
                                    "type": "string"
                                }
                            }
                        },
                        "secondaryPhone": {
                            "type": "string"
                        },
                        "mobilePhone": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        }
                    },
                    "required": [
                        "firstName",
                        "lastName",
                        "email",
                        "mobilePhone"
                    ]
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "source": {
                    "type": "string",
                    "x-input": "select"
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "x-input": "date-time"
                },
                "sourceNotes": {
                    "type": "string",
                    "x-input": "textArea"
                },
                "birthDate": {
                    "type": "string",
                    "format": "date-time",
                    "x-input": "date-time"
                },
                "archived": {
                    "type": "boolean"
                }
            },
            "required": [
                "startDate"
            ]
        },
...
        "standardSuccessResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
        "standardFailureResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "error": {
                    "type": "boolean"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "title": "errorMessage",
                        "properties": {
                            "field": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "updateClientInfo": {
            "title": "clientInfoRequest",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "updateClientSource": {
            "title": "updateClientSource",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "source": {
                    "type": "string"
                },
                "sourceName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "updateClientContact": {
            "title": "updateClientContact",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "secondaryPhone": {
                    "type": "string"
                },
                "mobilePhone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "updateClientAddress": {
            "title": "updateClientAddress",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "street1": {
                    "type": "string"
                },
                "street2": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
...
        "archiveClient": {
            "title": "archiveClient",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "archived": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
    "securityDefinitions": {},
...
}